requirements.txt
Flask
requests
beautifulsoup4
python-docx
PyPDF2